GasOps_weld_backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.5
requests-pkcs12==1.24
pydantic==2.11.7
python-multipart==0.0.12
python-dotenv==1.0.0
azure-ai-documentintelligence==1.0.0
azure-core==1.30.2
openai==1.104.2